+++ PocketShop/ViewModels/Common/RegisterViewModel.swift
