+++ PocketShop/Views/CustomerViews/FrontPage/ShopsScrollView.swift
